feat: qube list tab loading state
- create connector and vm for qube list tab
- add is loading parameter in qube list tab
- wrap qube list tab with animated opacity and lower if loading
- disable qube list tab if loading
