+++ lib/qube_list/qube_list_page.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:qube_project/database/database.dart';
+import 'package:qube_project/qube_list/qube_list_tab/qube_list_tab_connector.dart';
 import 'package:qube_project/qube_list/step_2/step_2_tab_connector.dart';
 import 'package:qube_project/qube_list/widgets/qube_list.dart';
-import 'package:qube_project/qube_list/widgets/qube_list_tab.dart';
+import 'package:qube_project/qube_list/qube_list_tab/qube_list_tab.dart';
 import 'package:qube_project/utils/const.dart';
 import 'package:qube_project/widgets/spacings.dart';
 
@@ -43,7 +44,7 @@ class _QubeListPageState extends State<QubeListPage> with SingleTickerProviderSt
     return Expanded(
       child: Column(
         children: [
-          QubeListTab(tabController: _tabController),
+          QubeListTabConnector(tabController: _tabController),
           const VerticalSpace(space: 32.0),
           Expanded(
             child: TabBarView(
